refactor: Tags 내부로 옮기기
