+++ src/app/(whitebg)/minari/Minari.module.css
@@ -21,31 +21,6 @@
   background: var(--color-bg-tertiary);
 }
 
-.keyword__wrapper {
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-  margin-top: 40px;
-}
-
-.keyword__container {
-  display: flex;
-  gap: 16px;
-}
-.keyword {
-  border-radius: var(--border-radius-rounded);
-  background: var(--black);
-  padding: 8px 16px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-bottom: 64px;
-}
-
-.keyword > span {
-  color: var(--white);
-}
-
 .button__wrapper {
   display: flex;
   justify-content: center;
