+++ src/app/(whitebg)/minari/_components/Tags/Tags.module.css
@@ -1,0 +1,25 @@
+.keyword__container {
+  display: flex;
+  gap: 16px;
+}
+
+.keyword {
+  border-radius: var(--border-radius-rounded);
+  background: var(--black);
+  padding: 8px 16px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-bottom: 64px;
+}
+
+.keyword__wrapper {
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+  margin-top: 40px;
+}
+
+.keyword > span {
+  color: var(--white);
+}
